Make Unpack work when packed and unpacked elements are the same size.
Previously, it did not work correctly in this case (which applies to arrays of any type other than char or boolean).  It would fail to unpack the array elements and instead would corrupt the first element of the packed array.  This occurred because it was calling ~Mul2 where it should be using ~Mul4, and these have different calling conventions which left the parameters set up incorrectly for the subsequent call to ~LongMove.

This bug was detected using the "Pascal acceptance test" distributed with Pascal-P5:
https://sourceforge.net/p/pascalp5/code/ci/master/tree/standard_tests/iso7185pat.pas

Here is an example program affected by this:

program UnpackTest(output);

var
        a: array[2..3] of integer;
        pa: packed array [4..5] of integer;

begin
pa[4] := 123;
pa[5] := 456;
unpack(pa, a, 2);
writeln(a[2]:1, a[3]:4, pa[4]:4, pa[5]:4); {should be 123 456 123 456}
end.